projects/projects/migrations/0017_auto_20141231_0933.py
# -*- coding: utf-8 -*-
from __future__ import unicode_literals

from django.db import models, migrations
import mptt.fields


class Migration(migrations.Migration):

    dependencies = [
        ('projects', '0016_auto_20141231_0932'),
    ]

    operations = [
        migrations.AddField(
            model_name='component',
            name='level',
            field=models.PositiveIntegerField(default=0, editable=False, db_index=True),
            preserve_default=False,
        ),
        migrations.AddField(
            model_name='component',
            name='lft',
            field=models.PositiveIntegerField(default=0, editable=False, db_index=True),
            preserve_default=False,
        ),
        migrations.AddField(
            model_name='component',
            name='parent',
            field=mptt.fields.TreeForeignKey(related_name='children', blank=True, to='projects.Component', null=True),
            preserve_default=True,
        ),
        migrations.AddField(
            model_name='component',
            name='rght',
            field=models.PositiveIntegerField(default=0, editable=False, db_index=True),
            preserve_default=False,
        ),
        migrations.AddField(
            model_name='component',
            name='tree_id',
            field=models.PositiveIntegerField(default=0, editable=False, db_index=True),
            preserve_default=False,
        ),
    ]
