requirements.txt
Django>=1.7
django-mptt==0.6.1
django-mptt-admin==0.1.10
